Reworked search to use python.
Instead of using Popen with grep, now I am searching files with
python entirely. this makes it behave much more robustly, but it
is somewhat slower. I don't think the slowdown is significant
enough to warrant using grep. HOWEVER, I left the grep
implementation, and switching to it is a boolean away ;).

I also fixed removed rofi_warn() in favor of the method:
QnAppRF.show_warning(). This works for both rofi and fzf!

Finally, I added an is_empty() method to the FileRepo class.

Testing seems to be OK. More reliable than before.